refactor: moved menu from header to components
